+++ src/main/kotlin/team/gsm/flooding/domain/auth/controller/AuthController.kt
@@ -2,25 +2,20 @@ package team.gsm.flooding.domain.auth.controller
 
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
-import team.gsm.flooding.domain.auth.dto.request.RegenerateEmailCodeRequest
 import team.gsm.flooding.domain.auth.dto.request.SignInRequest
 import team.gsm.flooding.domain.auth.dto.request.SignUpRequest
 import team.gsm.flooding.domain.auth.dto.response.ReissueTokenResponse
 import team.gsm.flooding.domain.auth.dto.response.SignInResponse
 import team.gsm.flooding.domain.auth.usecase.LogoutUsecase
-import team.gsm.flooding.domain.auth.usecase.RegenerateEmailCodeUsecase
 import team.gsm.flooding.domain.auth.usecase.ReissueTokenUsecase
 import team.gsm.flooding.domain.auth.usecase.SignInUsecase
 import team.gsm.flooding.domain.auth.usecase.SignUpUsecase
-import team.gsm.flooding.domain.auth.usecase.VerifyEmailUsecase
 
 @RestController
 @RequestMapping("auth")
@@ -28,8 +23,6 @@ class AuthController(
 	private val signUpUsecase: SignUpUsecase,
 	private val signInUsecase: SignInUsecase,
 	private val reissueTokenUsecase: ReissueTokenUsecase,
-	private val verifyEmailUsecase: VerifyEmailUsecase,
-	private val regenerateEmailCodeUsecase: RegenerateEmailCodeUsecase,
 	private val logoutUsecase: LogoutUsecase,
 ) {
 	@PostMapping("sign-up")
@@ -67,21 +60,4 @@ class AuthController(
 			ResponseEntity.ok(it)
 		}
 	}
-
-	@GetMapping("verify")
-	fun verifyEmail(
-		@RequestParam("email") email: String,
-		@RequestParam("code") code: String,
-	): ResponseEntity<String> {
-		verifyEmailUsecase.execute(email, code)
-		return ResponseEntity.ok("인증에 성공하였습니다.")
-	}
-
-	@PatchMapping("re-verify")
-	fun regenerateVerifyCode(
-		@RequestBody regenerateEmailCodeRequest: RegenerateEmailCodeRequest,
-	): ResponseEntity<Unit> =
-		regenerateEmailCodeUsecase.execute(regenerateEmailCodeRequest.email).let {
-			ResponseEntity.ok().build()
-		}
 }
