+++ src/main/resources/application-dev.yml
@@ -1,6 +1,7 @@
 server:
     servlet:
         context-path: /v1
+    base-url: "http://localhost:8080/v1"
 
 springdoc:
     show-actuator: true
